Fixed bug in win condition, moved win file & added namespaceMenu
