+++ intertwined/Assets/Scenes/ParkLevel.unity
@@ -46941,7 +46941,7 @@ GameObject:
   - component: {fileID: 1776352468}
   - component: {fileID: 1776352467}
   m_Layer: 0
-  m_Name: GameObject
+  m_Name: EndgameTarget
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
